ISENSIOT-SmartBabyMonitoring-main/requirements.txt
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
firebase_admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
influxdb_client==1.43.0
msgpack==1.0.8
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
reactivex==4.0.4
requests==2.32.3
RPi.GPIO==0.7.1
rsa==4.9
six==1.16.0
timeloop==1.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
